requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
pillow>=10.0.1
pytesseract==0.3.10
yt-dlp==2023.12.30